Added activity scope. Changed dagger implementation. Added subcomponents
