+++ app/build.gradle
@@ -67,8 +67,8 @@ dependencies {
     compile 'com.google.code.gson:gson:2.3.1'
     compile 'javax.inject:javax.inject:1'
 
-    compile 'com.google.dagger:dagger:2.0'
-    apt 'com.google.dagger:dagger-compiler:2.0'
+    compile 'com.google.dagger:dagger:2.0.1'
+    apt 'com.google.dagger:dagger-compiler:2.0.1'
     compile 'org.glassfish:javax.annotation:10.0-b28'
 
     testCompile 'org.robolectric:robolectric:3.0'
@@ -86,7 +86,7 @@ dependencies {
     testCompile 'com.squareup.okhttp:mockwebserver:2.5.0'
 
     // instrumentation tests
-    androidTestApt 'com.google.dagger:dagger-compiler:2.0'
+    androidTestApt 'com.google.dagger:dagger-compiler:2.0.1'
     androidTestCompile 'com.android.support.test.espresso:espresso-core:2.2.1'
     androidTestCompile 'com.android.support.test.espresso:espresso-intents:2.2.1'
     androidTestCompile 'com.android.support.test:runner:0.4.1'
