+++ app/src/main/java/co/infinum/pokemon/activities/PokemonDetailsActivity.java
@@ -7,9 +7,8 @@ import javax.inject.Inject;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
+import co.infinum.pokemon.PokemonApp;
 import co.infinum.pokemon.R;
-import co.infinum.pokemon.dagger.components.DaggerPokemonDetailsComponent;
-import co.infinum.pokemon.dagger.components.PokemonDetailsComponent;
 import co.infinum.pokemon.dagger.modules.PokemonDetailsModule;
 import co.infinum.pokemon.models.Pokemon;
 import co.infinum.pokemon.mvp.presenters.PokemonDetailsPresenter;
@@ -47,10 +46,8 @@ public class PokemonDetailsActivity extends BaseActivity implements PokemonDetai
         ButterKnife.inject(this);
 
         Pokemon pokemon = (Pokemon) getIntent().getSerializableExtra(EXTRA_POKEMON);
-        PokemonDetailsComponent component = DaggerPokemonDetailsComponent.builder()
-                .pokemonDetailsModule(new PokemonDetailsModule(this))
-                .build();
-        component.inject(this);
+
+        PokemonApp.getInstance().getApplicationComponent().plus(new PokemonDetailsModule(this)).inject(this);
 
         pokemonDetailsPresenter.loadDetails(pokemon);
     }
