+++ app/src/main/java/co/infinum/pokemon/activities/PokemonListActivity.java
@@ -11,10 +11,9 @@ import javax.inject.Inject;
 
 import butterknife.ButterKnife;
 import butterknife.InjectView;
+import co.infinum.pokemon.PokemonApp;
 import co.infinum.pokemon.R;
 import co.infinum.pokemon.adapters.PokemonAdapter;
-import co.infinum.pokemon.dagger.components.DaggerPokemonListComponent;
-import co.infinum.pokemon.dagger.components.PokemonListComponent;
 import co.infinum.pokemon.dagger.modules.PokemonListModule;
 import co.infinum.pokemon.models.Pokemon;
 import co.infinum.pokemon.mvp.presenters.PokemonListPresenter;
@@ -37,10 +36,7 @@ public class PokemonListActivity extends BaseActivity implements PokemonListView
         pokemonListRecycler.setHasFixedSize(true);
         pokemonListRecycler.setLayoutManager(new LinearLayoutManager(this));
 
-        PokemonListComponent component = DaggerPokemonListComponent.builder()
-                .pokemonListModule(new PokemonListModule(this))
-                .build();
-        component.inject(this);
+        PokemonApp.getInstance().getApplicationComponent().plus(new PokemonListModule(this)).inject(this);
 
         pokemonListPresenter.loadPokemonList();
     }
