+++ app/src/main/java/co/infinum/pokemon/dagger/components/AppComponent.java
@@ -3,21 +3,28 @@ package co.infinum.pokemon.dagger.components;
 import javax.inject.Singleton;
 
 import co.infinum.pokemon.PokemonApp;
-import dagger.Component;
-import co.infinum.pokemon.dagger.modules.ApiModule;
 import co.infinum.pokemon.dagger.modules.DefaultExecutorsModule;
 import co.infinum.pokemon.dagger.modules.HostModule;
+import co.infinum.pokemon.dagger.modules.NetworkModule;
+import co.infinum.pokemon.dagger.modules.PokemonDetailsModule;
+import co.infinum.pokemon.dagger.modules.PokemonListModule;
+import dagger.Component;
 
 /**
  * Created by dino on 13/10/15.
  */
 @Component(modules = {
+        NetworkModule.class,
         HostModule.class,
         DefaultExecutorsModule.class,
-        ApiModule.class
 })
 @Singleton
 public interface AppComponent {
 
     void inject(PokemonApp app);
+
+    PokemonListComponent plus(PokemonListModule pokemonListModule);
+
+    PokemonDetailsComponent plus(PokemonDetailsModule pokemonDetailsModule);
+
 }
