+++ app/src/main/java/co/infinum/pokemon/dagger/scopes/ActivityScope.java
@@ -1,0 +1,16 @@
+package co.infinum.pokemon.dagger.scopes;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+import javax.inject.Scope;
+
+/**
+ * @author Koc
+ *         ivan.kocijan@infinum.hr
+ * @since 21/11/15
+ */
+@Scope
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ActivityScope {
+}
