+++ app/src/test/java/co/infinum/pokemon/dagger/components/AppTestComponent.java
@@ -6,7 +6,6 @@ import co.infinum.pokemon.PokemonApp;
 import co.infinum.pokemon.dagger.modules.MockHostModule;
 import co.infinum.pokemon.dagger.modules.SynchronousExecutorsModule;
 import dagger.Component;
-import co.infinum.pokemon.dagger.modules.ApiModule;
 
 /**
  * Created by dino on 13/10/15.
